+++ example/App.tsx
@@ -15,12 +15,17 @@ const data: Data[] = [
 
 function App(): React.JSX.Element {
   const [item, setItem] = React.useState<string | JSX.Element>('');
-  const onDataSelect = (e: Data): void => setItem(e.label);
+  const onDataSelect = (data: Data) =>
+    setItem(data.label);
+  const onMultiDataSelect = (datas: Data[]) =>
+    setItem(datas.map((data: Data) =>
+      data.label
+    ).join(", "));
 
   return (
     <>
       <View style={{ height: 40 }} />
-      <ScrollView>
+      <ScrollView style={{ paddingHorizontal: 8 }}>
         <View style={{ height: 40 }} />
         <Select data={data} onSelect={onDataSelect} />
         <Text>Selected: {item || 'None'} (scroll down)</Text>
@@ -32,27 +37,209 @@ function App(): React.JSX.Element {
           }}
         >
           The dropdown menu will display above the input box when there
-          isn&quot;t enough space below
+          isn&apos;t enough space below
         </Text>
         <Select
           data={data}
-          onSelect={(e: Data) => {
-            console.log(e.label);
-          }}
+          onSelect={onDataSelect}
           placeholderText="Select an item"
           boxStyle={{
-            alignSelf: 'center',
-            width: 200,
-          }}
-          listStyle={{
             alignSelf: 'center',
             width: 200,
           }}
         />
         <MultiSelect
           data={data}
-          onSelect={(e: Data[]) => console.log(e)}
-          selectedItemStyle={{ backgroundColor: 'red' }}
+          onSelect={onMultiDataSelect}
+        />
+        <View style={{ height: 400 }}/>
+        <Text>Single Selects:</Text>
+        <View style={{ flexDirection: 'row', height: 100 }}>
+          <View style={{ flex: 1 }}>
+            <Select
+              data={data}
+              onSelect={onDataSelect}
+            />
+          </View>
+          <View style={{ flex: 1 }}>
+            <Select
+              data={data}
+              onSelect={onDataSelect}
+            />
+          </View>
+          <View style={{ flex: 1 }}>
+            <Select
+              data={data}
+              onSelect={onDataSelect}
+            />
+          </View>
+        </View>
+        <Text>Multi Selects:</Text>
+        <View style={{ height: 400 }}>
+          <View style={{ flexDirection: 'row' }}>
+            <View style={{ flex: 1 }}>
+              <MultiSelect
+                data={data}
+                onSelect={onMultiDataSelect}
+              />
+            </View>
+            <View style={{ flex: 1 }}>
+              <MultiSelect
+                data={data}
+                onSelect={onMultiDataSelect}
+              />
+            </View>
+            <View style={{ flex: 1 }}>
+              <MultiSelect
+                data={data}
+                onSelect={onMultiDataSelect}
+              />
+            </View>
+          </View>
+          <Text style={{ textAlign: 'center' }}>
+            Select more than one item and see me move!
+          </Text>
+        </View>
+        <Text>Styled Single Select:</Text>
+        <Select
+          data={data}
+          onSelect={onDataSelect}
+          defaultValue={data[0]}
+          listHeight={300}
+          placeholderText='I am very stylish'
+          boxStyle={{
+            alignSelf: 'center',
+            backgroundColor: '#48c',
+            borderTopRightRadius: 34,
+            borderBottomRightRadius: 16,
+            borderTopLeftRadius: 16,
+            borderBottomLeftRadius: 34,
+            borderColor: 'darkgreen',
+            borderLeftWidth: 2,
+            borderRightWidth: 2,
+            borderTopWidth: 1,
+            borderBottomWidth: 4,
+            width: 300,
+          }}
+          boxTextStyle={{
+            color: 'yellow',
+            fontFamily: 'courier new',
+            fontSize: 24,
+            fontStyle: 'italic',
+            fontWeight: 'bold',
+            paddingHorizontal: 8,
+            textShadowColor: 'brown',
+            textShadowOffset: { width: -2, height: 2 },
+            textShadowRadius: 4,
+          }}
+          listStyle={{
+            backgroundColor: '#abc',
+            borderTopRightRadius: 16,
+            borderBottomRightRadius: 34,
+            borderTopLeftRadius: 34,
+            borderBottomLeftRadius: 16,
+            borderColor: 'darkgreen',
+            borderLeftWidth: 4,
+            borderRightWidth: 4,
+            borderTopWidth: 8,
+            borderBottomWidth: 2,
+            width: '50%',
+          }}
+          listTextStyle={{
+            alignSelf: 'center',
+            color: 'brown',
+            fontFamily: 'times new roman',
+            fontSize: 20,
+          }}
+          selectedItemStyle={{
+            backgroundColor: 'yellow',
+            borderTopColor: 'darkgreen',
+            borderBottomColor: 'darkgreen',
+            borderTopWidth: 2,
+            borderBottomWidth: 2,
+          }}
+          dropdownArrowColor='darkgreen'
+        />
+        <View style={{height: 50}} />
+        <Text>Styled Multi Select:</Text>
+        <MultiSelect
+          data={data}
+          onSelect={onMultiDataSelect}
+          defaultValue={data}
+          listHeight={300}
+          placeholderText='I am very stylish'
+          boxStyle={{
+            alignSelf: 'center',
+            backgroundColor: '#48c',
+            borderTopRightRadius: 34,
+            borderBottomRightRadius: 16,
+            borderTopLeftRadius: 16,
+            borderBottomLeftRadius: 34,
+            borderColor: 'darkgreen',
+            borderLeftWidth: 2,
+            borderRightWidth: 2,
+            borderTopWidth: 1,
+            borderBottomWidth: 4,
+            width: 300,
+          }}
+          boxTextStyle={{
+            color: 'yellow',
+            fontFamily: 'courier new',
+            fontSize: 24,
+            fontStyle: 'italic',
+            fontWeight: 'bold',
+            paddingHorizontal: 8,
+            textShadowColor: 'brown',
+            textShadowOffset: { width: -2, height: 2 },
+            textShadowRadius: 4,
+          }}
+          boxTextHighlightStyle={{
+            backgroundColor: 'darkgreen',
+            borderColor: '#abc',
+            borderLeftWidth: 4,
+            borderRightWidth: 4,
+            borderTopLeftRadius: 10,
+            borderTopRightRadius: 20,
+            borderBottomLeftRadius: 20,
+            borderBottomRightRadius: 10,
+            paddingHorizontal: 2,
+          }}
+          listStyle={{
+            backgroundColor: '#abc',
+            borderTopRightRadius: 16,
+            borderBottomRightRadius: 34,
+            borderTopLeftRadius: 34,
+            borderBottomLeftRadius: 16,
+            borderColor: 'darkgreen',
+            borderLeftWidth: 4,
+            borderRightWidth: 4,
+            borderTopWidth: 8,
+            borderBottomWidth: 2,
+            width: '50%',
+          }}
+          listTextStyle={{
+            alignSelf: 'center',
+            color: 'brown',
+            fontFamily: 'times new roman',
+            fontSize: 20,
+          }}
+          selectedItemStyle={{
+            backgroundColor: 'yellow',
+            borderTopColor: 'darkgreen',
+            borderBottomColor: 'darkgreen',
+            borderTopWidth: 2,
+            borderBottomWidth: 2,
+          }}
+          dropdownArrowColor='darkgreen'
+          clearButtonStyle={{
+            backgroundColor: 'darkgreen',
+            borderTopLeftRadius: 34,
+            borderBottomRightRadius: 34,
+            borderTopRightRadius: 6,
+            borderBottomLeftRadius: 6,
+            borderColor: 'darkgreen'
+          }}
+          clearButtonIconColor='#abc'
         />
         <View style={{ height: 700 }} />
       </ScrollView>